Show all noun LU numbers
So far, not used noun LUs (i.e. those not participating in any LVC) were
ignored and the rest was renumbered. Thus if the second and the third LU
out of four were used, the page showed them as LU1 and LU2.

Now all are displayed, but LU1 and LU4 are empty, just with a notice
that this LU does not participate in any LVC.